+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-09-01 03:13:23
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 22ec26b7b3545ae2de88e34ce78d4c17f051a778
+
+## [ForwardWidgets-2kuai] - 2025-09-01 03:13:23
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: ac69444011b3ebb19d2501a760973398996e628c
+
+## [ForwardWidgets-pack1r] - 2025-09-01 03:13:23
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: fd531b2a5e53ae2b23f909e5922180d4da4da68b
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-31 03:00:16
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main